task_7_1/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
